Refactor - implement Atomic Design
