+++ src/components/atoms/IconSearch/IconSearch.tsx
@@ -1,0 +1,12 @@
+import SearchIcon from '@mui/icons-material/Search';
+
+const searchIconStyles: React.CSSProperties = {
+  position: 'absolute',
+  left: '0.5rem',
+};
+
+const IconSearch = () => {
+  return <SearchIcon style={searchIconStyles} />;
+};
+
+export default IconSearch;
