+++ src/components/atoms/Logo/Logo.tsx
@@ -1,0 +1,11 @@
+import Jrr from '../../../assets/vector/jrrt_elipse.svg?react';
+
+const logoStyles = {
+  maxHeight: '130px',
+  maxWidth: '130px',
+};
+
+const Logo = () => {
+  return <Jrr style={logoStyles} />;
+};
+export default Logo;
