+++ src/components/molecules/SearchInput/SearchInput.tsx
@@ -1,15 +1,11 @@
 import { Box, TextField } from '@mui/material';
-import SearchIcon from '@mui/icons-material/Search';
-import {
-  searchIconStyles,
-  searchInputStyles,
-  searchTextFieldStyles,
-} from './StyledSearchIcon.style.ts';
+import { searchInputStyles, searchTextFieldStyles } from './StyledSearchIcon.style.ts';
+import IconSearch from '../../atoms/IconSearch/IconSearch.tsx';
 
 const SearchInput = () => {
   return (
     <Box sx={searchInputStyles}>
-      <SearchIcon style={searchIconStyles} />
+      <IconSearch />
       <TextField id="input-search" label="Szukaj..." sx={searchTextFieldStyles} />
     </Box>
   );
